backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
google-generativeai>=0.7.0
qdrant-client>=1.9.0
python-dotenv>=1.0.0
pydantic>=2.0.0
httpx>=0.27.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
